add new hookpoint
