+++ event.yml
@@ -1,9 +1,19 @@
+eccube.event.route.admin_product_product_new.response:
+    - [onResponseAdminProductNewBefore, NORMAL]
+eccube.event.route.admin_product_product_edit.response:
+    - [onResponseAdminProductEditBefore, NORMAL]
+admin.product.edit.complete:
+    - [onResponseAdminProductEditAfter, NORMAL]
+eccube.event.route.product_detail.response:
+    - [onResponseProductsDetailBefore, NORMAL]
+
+# deprecated event (for v3.0.0 - 3.0.8)
 eccube.event.render.admin_product_product_new.before:
     - [onRenderAdminProductNewBefore, NORMAL]
 eccube.event.render.admin_product_product_edit.before:
     - [onRenderAdminProductEditBefore, NORMAL]
 eccube.event.controller.admin_product_product_edit.after:
     - [onAdminProductEditAfter, NORMAL]
-
 eccube.event.render.product_detail.before:
     - [onRenderProductsDetailBefore, NORMAL]
+
